// src/common/workflow/accion-auditoria.enum.ts
export enum AccionAuditoria {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  DELETE_LOGICAL = 'DELETE_LOGICAL',
  SUBMIT_FOR_REVIEW = 'SUBMIT_FOR_REVIEW',
  APPROVE = 'APPROVE',
  REJECT = 'REJECT',
  RESUBMIT = 'RESUBMIT',
  GENERATE_DOCUMENT = 'GENERATE_DOCUMENT',
  CREATE_VERSION = 'CREATE_VERSION',
  SYNC = 'SYNC',
}